refactor(cli): move detail reports to CLI
